Added ECS Draft

+++ pde-engine/Headers/pde/Util/hash.h
@@ -1,8 +1,8 @@
 #pragma once
 
-const inline unsigned long hash(const char* str) {
-    unsigned long hash = 5381;
-    int c;
+const inline uint64_t hash(const char* str) {
+    uint64_t hash = 5381;
+    uint64_t c;
 
     while (c = *str++)
         hash = ((hash << 5) + hash) + c; /* hash * 33 + c */
